[TEST] 카테고리 서비스 테스트 수정

+++ src/main/java/stitch/crew/hour/category/service/CategoryService.java
@@ -61,6 +61,6 @@ public class CategoryService {
         Category category = categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.CATEGORY_NOT_FOUND));
         categoryRepository.delete(category);
-        imageService.deleteThumbnail(category.getThumbnail());
+        if (category.getThumbnail()!=null) imageService.deleteThumbnail(category.getThumbnail());
     }
 }
